requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
discord.py==2.5.2
-e git+https://github.com/Lu4head/MeuBotDiscord.git@7832e020b0b2269b10b6ada856cca2ee41ad5909#egg=discord_bot
frozenlist==1.6.0
ghp-import==2.1.0
google-auth==2.40.2
google-genai==1.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Markdown==3.8
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
multidict==6.4.4
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.3
rsa==4.9.1
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.20.0
